+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 Changelog
 =========
 
+### 1.? (In progress)
+- Added an example for setting YAML as default CLI output format
+
 ### 1.1.0
 - Updated input formatter to be compatible with hug 2.1.0 specification